+++ examples/envman/src/cmds/env_clean_cmd.rs
@@ -6,7 +6,7 @@ use peace::{
     rt_model::{outcomes::CmdOutcome, output::OutputWrite},
 };
 
-use crate::{cmds::EnvCmd, model::AppCycleError};
+use crate::{cmds::EnvCmd, model::EnvManError};
 
 /// Cleans up (deletes) the environment.
 #[derive(Debug)]
@@ -21,9 +21,9 @@ impl EnvCleanCmd {
     /// * `slug`: Username and repository of the application to download.
     /// * `version`: Version of the application to download.
     /// * `url`: URL to override where to download the application from.
-    pub async fn run<O>(output: &mut O) -> Result<(), AppCycleError>
+    pub async fn run<O>(output: &mut O) -> Result<(), EnvManError>
     where
-        O: OutputWrite<AppCycleError> + Send,
+        O: OutputWrite<EnvManError> + Send,
     {
         let states_saved = EnvCmd::run(output, true, |ctx| {
             StatesSavedReadCmd::exec(ctx).boxed_local()
@@ -81,21 +81,21 @@ impl EnvCleanCmd {
                         for (item_spec_id, error) in errors.iter() {
                             #[rustfmt::skip]
                             let diagnostic: &dyn Diagnostic = match error {
-                                // AppCycleError::AppCycleUrlBuild { url_candidate, error } => todo!(),
-                                // AppCycleError::EnvTypeParseError(_) => todo!(),
-                                // AppCycleError::ProfileSwitchToNonExistent { profile_to_switch_to, app_name } => todo!(),
-                                // AppCycleError::ProfileToCreateExists { profile_to_create, app_name } => todo!(),
+                                // EnvManError::EnvManUrlBuild { url_candidate, error } => todo!(),
+                                // EnvManError::EnvTypeParseError(_) => todo!(),
+                                // EnvManError::ProfileSwitchToNonExistent { profile_to_switch_to, app_name } => todo!(),
+                                // EnvManError::ProfileToCreateExists { profile_to_create, app_name } => todo!(),
 
-                                AppCycleError::PeaceItemSpecFileDownload(e) => e,
-                                AppCycleError::PeaceItemSpecTarX(e) => e,
-                                AppCycleError::InstanceProfileItemSpec(e) => e,
-                                AppCycleError::IamPolicyItemSpec(e) => e,
-                                AppCycleError::IamRoleItemSpec(e) => e,
-                                AppCycleError::S3BucketItemSpec(e) => e,
-                                AppCycleError::S3ObjectItemSpec(e) => e,
-                                AppCycleError::PeaceRtError(e) => e,
-                                // AppCycleError::WouldCycleError(_) => todo!(),
-                                // AppCycleError::TokioRuntimeInit(_) => todo!(),
+                                EnvManError::PeaceItemSpecFileDownload(e) => e,
+                                EnvManError::PeaceItemSpecTarX(e) => e,
+                                EnvManError::InstanceProfileItemSpec(e) => e,
+                                EnvManError::IamPolicyItemSpec(e) => e,
+                                EnvManError::IamRoleItemSpec(e) => e,
+                                EnvManError::S3BucketItemSpec(e) => e,
+                                EnvManError::S3ObjectItemSpec(e) => e,
+                                EnvManError::PeaceRtError(e) => e,
+                                // EnvManError::WouldCycleError(_) => todo!(),
+                                // EnvManError::TokioRuntimeInit(_) => todo!(),
 
                                 _ => error,
                             };