+++ examples/envman/src/cmds/env_discover_cmd.rs
@@ -2,17 +2,17 @@ use futures::FutureExt;
 use peace::{
     cmd::scopes::SingleProfileSingleFlowView,
     fmt::presentable::{Heading, HeadingLevel, ListNumbered},
-    rt::cmds::sub::StatesDesiredReadCmd,
+    rt::cmds::StatesDiscoverCmd,
     rt_model::output::OutputWrite,
 };
 
-use crate::{cmds::EnvCmd, model::AppCycleError};
+use crate::{cmds::EnvCmd, model::EnvManError};
 
 /// Shows the desired state of the environment.
 #[derive(Debug)]
-pub struct EnvDesiredCmd;
+pub struct EnvDiscoverCmd;
 
-impl EnvDesiredCmd {
+impl EnvDiscoverCmd {
     /// Shows the desired state of the environment.
     ///
     /// # Parameters
@@ -21,16 +21,33 @@ impl EnvDesiredCmd {
     /// * `slug`: Username and repository of the application to download.
     /// * `version`: Version of the application to download.
     /// * `url`: URL to override where to download the application from.
-    pub async fn run<O>(output: &mut O) -> Result<(), AppCycleError>
+    pub async fn run<O>(output: &mut O) -> Result<(), EnvManError>
     where
-        O: OutputWrite<AppCycleError> + Send,
+        O: OutputWrite<EnvManError> + Send,
     {
         EnvCmd::run(output, true, |ctx| {
             async {
-                let states_desired = StatesDesiredReadCmd::exec(ctx).await?;
+                let (states_current, states_desired) =
+                    StatesDiscoverCmd::current_and_desired(ctx).await?;
+                let states_current_raw_map = &**states_current;
                 let states_desired_raw_map = &**states_desired;
 
                 let SingleProfileSingleFlowView { output, flow, .. } = ctx.view();
+                let states_current_presentables = {
+                    let states_current_presentables = flow
+                        .graph()
+                        .iter_insertion()
+                        .map(|item_spec| {
+                            let item_spec_id = item_spec.id();
+                            match states_current_raw_map.get(item_spec_id) {
+                                Some(state_current) => (item_spec_id, format!(": {state_current}")),
+                                None => (item_spec_id, String::from(": <unknown>")),
+                            }
+                        })
+                        .collect::<Vec<_>>();
+
+                    ListNumbered::new(states_current_presentables)
+                };
                 let states_desired_presentables = {
                     let states_desired_presentables = flow
                         .graph()
@@ -49,6 +66,9 @@ impl EnvDesiredCmd {
 
                 output
                     .present(&(
+                        Heading::new(HeadingLevel::Level1, "States Current"),
+                        states_current_presentables,
+                        "\n",
                         Heading::new(HeadingLevel::Level1, "States Desired"),
                         states_desired_presentables,
                         "\n",