+++ examples/envman/src/model.rs
@@ -1,7 +1,7 @@
 //! Data structures
 
 pub use self::{
-    app_cycle_error::AppCycleError, env_type::EnvType, env_type_parse_error::EnvTypeParseError,
+    env_type::EnvType, env_type_parse_error::EnvTypeParseError, envman_error::EnvManError,
     profile_switch::ProfileSwitch, repo_slug::RepoSlug, repo_slug_error::RepoSlugError,
     web_app_file_id::WebAppFileId,
 };
@@ -9,9 +9,9 @@ pub use self::{
 #[cfg(not(target_arch = "wasm32"))]
 pub mod cli_args;
 
-mod app_cycle_error;
 mod env_type;
 mod env_type_parse_error;
+mod envman_error;
 mod profile_switch;
 mod repo_slug;
 mod repo_slug_error;