+++ examples/envman/src/rt_model.rs
@@ -1,0 +1,5 @@
+//! Runtime data types for the EnvMan example
+
+pub use self::envman_cmd_ctx::EnvManCmdCtx;
+
+mod envman_cmd_ctx;